requirements-dev.txt
# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-xdist==3.3.1

# Code Quality
black==23.3.0
flake8==6.0.0
isort==5.12.0
mypy==1.3.0
pylint==2.17.4

# Documentation
sphinx==6.2.1
sphinx-rtd-theme==1.2.0
sphinx-autodoc-typehints==1.23.0

# Development Tools
pre-commit==3.3.2
ipython==8.12.0
jupyter==1.0.0
notebook==6.5.4

# Type Checking
types-PyYAML==6.0.12.9
types-requests==2.31.0.2
types-setuptools==67.8.0.0

# Debugging
debugpy==1.6.7
memory-profiler==0.61.0
line-profiler==4.1.1

# Version Control
gitpython==3.1.31

# Optional: GPU Profiling
py-spy==0.3.14
torch-tb-profiler==0.4.1

# Optional: Visualization
matplotlib==3.7.1
seaborn==0.12.2
tensorboard==2.13.0

# Optional: Data Processing
pandas==2.0.2
scikit-learn==1.2.2
scikit-image==0.20.0

# Optional: API Development
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8

# Optional: Security
bandit==1.7.5
safety==2.3.5 